Issue-545: WIP Create SearchClient component
